+++ SolrNet.Tests.Common/Mocks/MSolrBasicOperations.cs
@@ -81,6 +81,16 @@ namespace SolrNet.Tests.Mocks {
             return delete.Invoke(ids, q, parameters);
         }
 
+        public ResponseHeader AtomicUpdate(string uniqueKey, string id, IEnumerable<AtomicUpdateSpec> updateSpecs, AtomicUpdateParameters parameters)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<ResponseHeader> AtomicUpdateAsync(string uniqueKey, string id, IEnumerable<AtomicUpdateSpec> updateSpecs, AtomicUpdateParameters parameters)
+        {
+            throw new NotImplementedException();
+        }
+
         public Task<ResponseHeader> CommitAsync(CommitOptions options)
         {
             throw new NotImplementedException();